fix add_plot
